Updated GUI Main
Added test cases and comments
